Split kivy classes into separate files, added pop up for move not found and error code at backend
