+++ src/movesearch.kv
@@ -1,0 +1,86 @@
+#: include main.kv
+
+<MoveSearch>:
+    name: 'movesearch'
+
+    code: code
+    move_name: move_name
+    belt_colour: belt_colour
+    related_moves: related_moves
+    in_module: in_module
+    user_search: user_search
+
+    FloatLayout:
+        Label:
+            text: 'Get a move'
+            font_size: 30
+            font_size: (root.width**2 + root.height**2) / 14**4
+            pos_hint: {"x":0.32, "top":0.95}
+            size_hint: 0.35, 0.15
+
+        Label:
+            id: move_name
+            pos_hint: {"x":0.32, "top":0.85}
+            size_hint: 0.25, 0.1
+            text: "Name: "
+
+        Label:
+            id: code
+            pos_hint: {"x":0.32, "top":0.8}
+            size_hint: 0.25, 0.1
+            text: "Move ID: "
+
+        Label:
+            id: belt_colour
+            pos_hint: {"x":0.32, "top":0.75}
+            size_hint: 0.25, 0.1
+            text: "Belt: "
+
+        Label:
+            id: related_moves
+            pos_hint: {"x":0.32, "top":0.7}
+            size_hint: 0.25, 0.1
+            text: "Related: "
+
+        Label:
+            id:in_module
+            pos_hint: {"x":0.32, "top":0.65}
+            size_hint: 0.25, 0.1
+            text: "Module: "
+
+        Button:
+            text: 'Get move'
+            size_hint: 0.2, 0.1
+            font_size: (root.width**2 + root.height**2) / 14**4
+            pos_hint: {"x":0.35, "y":0.1}
+            on_release: root.get_move()
+
+        Button:
+            text: 'Back'
+            size_hint: 0.2, 0.1
+            font_size: (root.width**2 + root.height**2) / 14**4
+            on_release: app.root.current = 'reference'
+            pos_hint: {"x":0.1, "y":0.1}
+            on_release:
+                root.manager.transition.direction = "right"
+
+        Button:
+            text: 'Reset'
+            size_hint: 0.2, 0.1
+            font_size: (root.width**2 + root.height**2) / 14**4
+            pos_hint: {"x":0.6, "y":0.1}
+            on_release: root.clear_result()
+
+        TextInput:
+            id: user_search
+            font_size: (root.width**2 + root.height**2) / 14**4
+            multiline: False
+            size_hint: 0.4, 0.15
+            pos_hint: {"x":0.35, "y":0.3}
+
+        Button:
+            text: 'Search'
+            size_hint: 0.2, 0.1
+            font_size: (root.width**2 + root.height**2) / 14**4
+            pos_hint: {"x":0.35, "y":0.2}
+            on_release: root.search_by_name()
